Update 10-server-delivery.md
The example code would not run.  I found that the .js file will run if the "if socket recovered" statement was inside the "socket.on chat message" expression.

*** The resulting code does not produce the desired output!***

Also, the word "async" was added to the code, but was not highlighted in the tutorial, so that was changed as well.